requirements.txt
pytest>=7.4.0
pytest-html>=3.2.0
allure-pytest>=2.13.0
requests>=2.31.0
PyYAML>=6.0.1
pymysql>=1.1.0
redis>=5.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
